--- a/.github/workflows/train_workflow.yml
+++ b/.github/workflows/train_workflow.yml
@@ -19,24 +19,18 @@ jobs:
         with:
           python-version: '3.10'
 
-      - name: Cache PyTorch
-        uses: actions/cache@v3
-        with:
-          path: |
-            ~/.cache/pip
-            ${{ runner.tool_cache }}/Python
-          key: ${{ runner.os }}-pip-${{ hashFiles('requirements.txt') }}
-          restore-keys: |
-            ${{ runner.os }}-pip-
+      - name: Install uv
+        run: |
+          curl -LsSf https://astral.sh/uv/install.sh | sh
+          echo "$HOME/.cargo/bin" >> $GITHUB_PATH
 
       - name: Install dependencies
         run: |
-          pip install numpy torch
+          uv pip install numpy torch --no-cache
           
       - name: Create and run training script
         run: |
           echo "${{ inputs.script_content }}" > train.py
-          cat train.py  # Debug: print the content
           python train.py > training.log 2>&1
         
       - name: Upload logs