--- a/.github/workflows/train_workflow.yml
+++ b/.github/workflows/train_workflow.yml
@@ -24,13 +24,16 @@ jobs:
           curl -LsSf https://astral.sh/uv/install.sh | sh
           echo "$HOME/.cargo/bin" >> $GITHUB_PATH
 
-      - name: Install dependencies
+      - name: Create virtual environment and install dependencies
         run: |
+          uv venv
+          source .venv/bin/activate
           uv pip install numpy torch --no-cache
           
       - name: Create and run training script
         run: |
           echo "${{ inputs.script_content }}" > train.py
+          source .venv/bin/activate
           python train.py > training.log 2>&1
         
       - name: Upload logs