--- a/.github/workflows/train_workflow.yml
+++ b/.github/workflows/train_workflow.yml
@@ -18,17 +18,16 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
+          cache: 'pip'
 
       - name: Install uv
-        run: |
-          curl -LsSf https://astral.sh/uv/install.sh | sh
-          echo "$HOME/.cargo/bin" >> $GITHUB_PATH
+        uses: astral-sh/setup-uv@v1
 
       - name: Create virtual environment and install dependencies
         run: |
           uv venv
           source .venv/bin/activate
-          uv pip install numpy torch --no-cache
+          uv pip install --no-deps numpy torch
           
       - name: Create and run training script
         run: |