--- a/.github/workflows/train_workflow.yml
+++ b/.github/workflows/train_workflow.yml
@@ -18,10 +18,19 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
-          cache: 'pip'
 
       - name: Install uv
         uses: astral-sh/setup-uv@v1
+
+      - name: Cache UV packages
+        uses: actions/cache@v3
+        with:
+          path: |
+            ~/.cache/uv
+            .venv
+          key: ${{ runner.os }}-uv-numpy-torch-v1
+          restore-keys: |
+            ${{ runner.os }}-uv-
 
       - name: Create virtual environment and install dependencies
         run: |