--- a/.github/workflows/train_workflow.yml
+++ b/.github/workflows/train_workflow.yml
@@ -42,8 +42,10 @@ jobs:
         run: |
           echo "${{ inputs.script_content }}" > train.py
           source .venv/bin/activate
-          python train.py > training.log 2>&1
-        
+          # python train.py > training.log 2>&1
+          cat train.py
+          python train.py
+          
       - name: Upload logs
         uses: actions/upload-artifact@v3
         if: always()