--- a/requirements.txt
+++ b/requirements.txt
@@ -1,6 +1,6 @@
 PyGithub
 aiohttp
 discord.py
-audioop-lts # discord.py imports using * syntax 
 python-dotenv
 requests
+pytest