--- a/torch/train.py
+++ b/torch/train.py
@@ -1,0 +1,6 @@
+import torch
+
+x = torch.randn(2, 3)
+y = torch.randn(3, 2)
+z = torch.matmul(x, y)
+print(z)