re: 

https://github.com/gpu-mode/discord-cluster-manager/issues/9

Results after run locally (macbook pro m1)

```
==================================================
                BENCHMARK RESULTS                 
==================================================
Test Type                     Duration            
--------------------------------------------------
With Cache                    26.76s              
Without Cache                 68.07s              
With UV                       9.28s               
Cache Speedup                 41.31s              
UV vs Cache Speedup           17.48s              
==================================================

Test Output:
--------------------------------------------------
Matrix multiplication result:
 [[19 22]
 [43 50]]```